Разобьёт код на модули
